Allow upgrading from HTTP/1.1 to other protocols (#481)
This patch implements the http_upgrade_request_protocols configuration
parameter to controls client-initiated and server-confirmed switching
from  HTTP to another protocol (or to several protocols) using HTTP
Upgrade mechanism (RFC 7230 Section 6.7).

Squid itself does not understand the protocols being upgraded to and
participates in the upgraded communication only as a dumb TCP proxy.

This is a Measurement Factory project